requirements_py311.txt
annotated-types==0.7.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
contourpy==1.3.2
cryptography==45.0.6
cycler==0.12.1
dill==0.4.0
filelock==3.19.1
fonttools==4.59.1
fsspec==2025.7.0
ibm-cloud-sdk-core==3.24.2
ibm-platform-services==0.68.1
idna==3.10
Jinja2==3.1.6
joblib==1.5.1
kiwisolver==1.4.9
MarkupSafe==3.0.2
matplotlib==3.10.5
mpmath==1.3.0
networkx==3.5
numpy==1.23.0
packaging==25.0
pandas==1.5.1
pbr==7.0.0
pillow==11.3.0
psutil==7.0.0
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
PyJWT==2.10.1
pyparsing==3.2.3
pyspnego==0.11.2
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
-e git+ssh://git@github.com/AnikaJerin/Quantum-Enhanced-Graph-Learning-for-Molecular-Design-in-Biomaterials.git@f163833ed917303c9725dcdb85ba6e326dd99204#egg=qegl_biomaterials
qiskit==2.1.2
qiskit-aer==0.17.1
qiskit-ibm-runtime==0.30.0
qiskit-machine-learning==0.8.2
rdkit-pypi==2022.9.5
requests==2.32.5
requests_ntlm==1.3.0
rustworkx==0.17.1
scikit-learn==1.7.1
scipy==1.13.1
six==1.17.0
stevedore==5.4.1
sympy==1.14.0
threadpoolctl==3.6.0
torch==2.8.0
torchaudio==2.8.0
torchvision==0.23.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
websocket-client==1.8.0
